Structures that have been manually edited and minimized in explicit solvent have been added, as has the script for running the minimizations in openmm
Why
---
The structures contained in /minimized/ where minimized in vaccuum (with crystallographic waters present), and no attempt was made to fix the structures
that failed minimization. A number of 'XXXX-fixed.pdb' structures contained mistakes and crystallographic artifacts that needed manually editing. All
details of the manual edits can be found in README.md, which is also added to this commit.